swe: move to new directory
