+++ case_studies/tohoku/hazard/run_continuous_adjoint.py
@@ -4,7 +4,7 @@ import argparse
 import os
 
 from adapt_utils.case_studies.tohoku.options.hazard_options import TohokuHazardOptions
-from adapt_utils.unsteady.swe.tsunami.solver import AdaptiveTsunamiProblem
+from adapt_utils.swe.tsunami.solver import AdaptiveTsunamiProblem
 
 
 # --- Parse arguments
