+++ case_studies/tohoku/hazard/run_qmesh_convergence.py
@@ -5,7 +5,7 @@ import os
 
 from adapt_utils.case_studies.tohoku.options.hazard_options import TohokuHazardOptions
 from adapt_utils.io import OuterLoopLogger
-from adapt_utils.unsteady.swe.tsunami.solver import AdaptiveTsunamiProblem
+from adapt_utils.swe.tsunami.solver import AdaptiveTsunamiProblem
 
 
 # --- Parse arguments
