+++ case_studies/tohoku/inversion/box/discrete.py
@@ -13,9 +13,9 @@ from adapt_utils.case_studies.tohoku.options.box_options import TohokuBoxBasisOp
 from adapt_utils.norms import total_variation, vecnorm
 from adapt_utils.optimisation import minimise_bfgs
 from adapt_utils.plotting import *
+from adapt_utils.swe.tsunami.conversion import lonlat_to_utm
 from adapt_utils.unsteady.solver import AdaptiveProblem
 from adapt_utils.unsteady.solver_adjoint import AdaptiveDiscreteAdjointProblem
-from adapt_utils.unsteady.swe.tsunami.conversion import lonlat_to_utm
 
 
 class DiscreteAdjointTsunamiProblem(AdaptiveDiscreteAdjointProblem):
