+++ case_studies/tohoku/inversion/plot_optimised_source.py
@@ -7,8 +7,8 @@ import os
 import sys
 
 from adapt_utils.plotting import *
+from adapt_utils.swe.tsunami.conversion import lonlat_to_utm
 from adapt_utils.unsteady.solver import AdaptiveProblem
-from adapt_utils.unsteady.swe.tsunami.conversion import lonlat_to_utm
 
 
 # --- Parse arguments
