+++ case_studies/tohoku/inversion/radial/discrete.py
@@ -15,7 +15,7 @@ from adapt_utils.optimisation import minimise_bfgs
 from adapt_utils.plotting import *
 from adapt_utils.unsteady.solver import AdaptiveProblem
 from adapt_utils.unsteady.solver_adjoint import AdaptiveDiscreteAdjointProblem
-from adapt_utils.unsteady.swe.tsunami.conversion import lonlat_to_utm
+from adapt_utils.swe.tsunami.conversion import lonlat_to_utm
 
 
 class DiscreteAdjointTsunamiProblem(AdaptiveDiscreteAdjointProblem):
