+++ case_studies/tohoku/options/options.py
@@ -4,8 +4,8 @@ import netCDF4
 import os
 
 from adapt_utils.io import load_mesh
-from adapt_utils.unsteady.swe.tsunami.options import TsunamiOptions
-from adapt_utils.unsteady.swe.tsunami.conversion import from_latlon
+from adapt_utils.swe.tsunami.options import TsunamiOptions
+from adapt_utils.swe.tsunami.conversion import from_latlon
 
 
 __all__ = ["TohokuOptions"]
