+++ swe/adjoint.py
