+++ swe/equation.py
