+++ swe/solver.py
