+++ unsteady/test_cases/spaceship/run_fixed_mesh.py
@@ -7,9 +7,9 @@ import os
 import sys
 from time import perf_counter
 
-from adapt_utils.unsteady.swe.turbine.solver import AdaptiveTurbineProblem
-from adapt_utils.unsteady.test_cases.spaceship.options import SpaceshipOptions
 from adapt_utils.plotting import *  # NOQA
+from adapt_utils.swe.turbine.solver import AdaptiveTurbineProblem
+from adapt_utils.unsteady.test_cases.spaceship.options import SpaceshipOptions
 
 
 # --- Parse arguments
