+++ unsteady/test_cases/turbine_array/plot_mesh_reynolds_number.py
@@ -5,8 +5,8 @@ import matplotlib.pyplot as plt
 import numpy as np
 
 from adapt_utils.plotting import *
+from adapt_utils.swe.turbine.solver import AdaptiveTurbineProblem
 from adapt_utils.unsteady.test_cases.turbine_array.options import TurbineArrayOptions
-from adapt_utils.unsteady.swe.turbine.solver import AdaptiveTurbineProblem
 
 
 # --- Parse arguments
