+++ unsteady/test_cases/turbine_array/run_continuous_adjoint.py
@@ -5,7 +5,7 @@ import numpy as np
 import os
 from time import perf_counter
 
-from adapt_utils.unsteady.swe.turbine.solver import AdaptiveTurbineProblem
+from adapt_utils.swe.turbine.solver import AdaptiveTurbineProblem
 from adapt_utils.unsteady.test_cases.turbine_array.options import TurbineArrayOptions
